+++ src/installers/url.rs
@@ -1,13 +1,12 @@
 // Internal module imports:
-// These `use` statements bring necessary types and functions into scope for this module.
-
 // `ToolEntry`: Represents a single tool's configuration as defined in your `tools.yaml` file.
 //              It's a struct that contains all possible configuration fields for a tool,
 //              such as name, version, source, URL, repository, etc.
 // `ToolState`: Represents the actual state of an *installed* tool. This struct is used to
 //              persist information about installed tools in the application's `state.json` file.
 //              It helps `setup-devbox` track what's installed, its version, and where it's located.
-use crate::schemas::sdb_schema::{ToolEntry, ToolState};
+use crate::schemas::state_file::ToolState;
+use crate::schemas::tools::ToolEntry;
 // Custom logging macros:
 // These macros (`log_debug!`, `log_error!`, `log_info!`, `log_warn!`) provide a
 // consistent and structured way to output messages at different severity levels.
@@ -377,5 +376,6 @@ pub fn install(tool_entry: &ToolEntry) -> Option<ToolState> {
         // Record any additional commands that were executed during installation.
         // This is useful for tracking what was done and potentially for cleanup during uninstall.
         additional_cmd_executed: tool_entry.additional_cmd.clone(),
+        configuration_manager: None,
     })
 }
